Move combine_dictionaries to extraction_utilities.py
